refactoring(chapter2/package): chapter2 패키지 도메인별로 분리
