--- code/src/main/java/com/study/objects/chapter2/domain/DiscountCondition.java
@@ -1,5 +1,0 @@
-package com.study.objects.chapter2.domain;
-
-public interface DiscountCondition {
-	boolean isSatisfiedBy(Screening screening);
-}
