+++ code/src/main/java/com/study/objects/chapter2/domain/common/Money.java
@@ -1,4 +1,4 @@
-package com.study.objects.chapter2.domain;
+package com.study.objects.chapter2.domain.common;
 
 import java.math.BigDecimal;
