+++ code/src/main/java/com/study/objects/chapter2/domain/discount/DiscountCondition.java
@@ -1,0 +1,7 @@
+package com.study.objects.chapter2.domain.discount;
+
+import com.study.objects.chapter2.domain.movie.Screening;
+
+public interface DiscountCondition {
+	boolean isSatisfiedBy(Screening screening);
+}
