+++ code/src/main/java/com/study/objects/chapter2/domain/discount/impl/PercentDiscountPolicy.java
@@ -1,4 +1,9 @@
-package com.study.objects.chapter2.domain;
+package com.study.objects.chapter2.domain.discount.impl;
+
+import com.study.objects.chapter2.domain.discount.DiscountCondition;
+import com.study.objects.chapter2.domain.common.Money;
+import com.study.objects.chapter2.domain.movie.Screening;
+import com.study.objects.chapter2.domain.discount.DiscountPolicy;
 
 public class PercentDiscountPolicy extends DiscountPolicy {
 	private double percent;
